refactor(data): move achivements to AchievementsData.ts and group data files under /data
